#125 Add IEnvironmentConfig.appBaseUrl.
