+++ web-frame/src/app-shared/config/defaults/environment.prod.ts
@@ -8,6 +8,7 @@ import { IEnvironmentConfig } from '../interfaces/environment-config.interface';
 /** @internal This variable should never be used directly; only by the build process. */
 export const ENVIRONMENT: Readonly<IEnvironmentConfig> = {
   apiBaseUrl: '/api/v1',
+  appBaseUrl: '/',
   appName: '',
   clientId: '',
   clientSecret: '',
