+++ web-frame/src/app-shared/config/defaults/environment.proxy.ts
@@ -8,6 +8,7 @@ import { IEnvironmentConfig } from '../interfaces/environment-config.interface';
 /** @internal This variable should never be used directly; only by the build process. */
 export const ENVIRONMENT: Readonly<IEnvironmentConfig> = {
   apiBaseUrl: '/api/v1',
+  appBaseUrl: '/',
   appName: 'WebFrame-3angleTECH',
   clientId: '3at-api',
   clientSecret: '5r5rd15c650f4849119eb894939d9fdaaf5f7d0e7e0f65de15b71bfa6411011y',
