+++ web-frame/src/app-shared/core/core.module.ts
@@ -19,7 +19,7 @@ import {
 } from './service/account/account.service';
 import { IApiEndpointBuilderService } from './service/api-endpoint-builder/api-endpoint-builder.interface';
 import { ApiEndpointBuilderService } from './service/api-endpoint-builder/api-endpoint-builder.service';
-import { APP_INITIALIZER_PROVIDERS } from './service/app-initializer.service';
+import { APP_INITIALIZER_PROVIDERS } from './service/app-initializer.factory';
 import {
   IJsonConverterService,
   JsonConverterService,
@@ -76,7 +76,6 @@ const SHARED_PIPES: Type<unknown>[] = [
 ];
 
 const SHARED_ROOT_PROVIDERS: Provider[] = [
-  APP_INITIALIZER_PROVIDERS,
   {
     provide: IAccountService,
     useClass: AccountService,
@@ -156,7 +155,10 @@ export class CoreModule {
   public static forRoot(): ModuleWithProviders<CoreModule> {
     return {
       ngModule: CoreModule,
-      providers: SHARED_ROOT_PROVIDERS,
+      providers: [
+        APP_INITIALIZER_PROVIDERS,
+        SHARED_ROOT_PROVIDERS,
+      ],
     };
   }
 }
